feat(networks): promote LayerNormNd to a public layer
Move the channels-first LayerNormNd out of the ConvNeXt net into
monai/networks/layers so any channels-first backbone can reuse it,
mirroring the DropPath layer. Export it from monai.networks.layers,
document it in the Layers docs, and add a dedicated layer test that
checks equivalence to nn.LayerNorm on the permuted channels-last tensor.
